+++ packages/amplify-category-auth/src/__tests__/commands/import.headless.test.ts
@@ -1,7 +1,14 @@
-import { executeAmplifyHeadlessCommand } from '../../../lib';
+import { executeAmplifyHeadlessCommand } from '../../../src';
 import { ImportAuthRequest } from 'amplify-headless-interface';
 import { messages } from '../../provider-utils/awscloudformation/assets/string-maps';
-import { stateManager } from 'amplify-cli-core';
+import { stateManager, FeatureFlags, JSONUtilities } from 'amplify-cli-core';
+import { printer } from 'amplify-prompts';
+
+jest.mock('amplify-prompts', () => ({
+  printer: {
+    info: jest.fn(),
+  },
+}));
 
 jest.mock('amplify-cli-core', () => ({
   stateManager: {
@@ -11,6 +18,12 @@ jest.mock('amplify-cli-core', () => ({
         awscloudformation: {},
       },
     }),
+  },
+  FeatureFlags: {
+    getBoolean: () => false,
+  },
+  JSONUtilities: {
+    parse: JSON.parse,
   },
 }));
 
@@ -127,7 +140,7 @@ describe('import auth headless', () => {
       },
     };
   });
-  
+
   beforeEach(() => {
     jest.clearAllMocks();
   });