+++ packages/amplify-cli/src/__tests__/extensions/amplify-helpers/get-project-meta.test.ts
@@ -30,10 +30,10 @@ describe('getProjectMeta', () => {
   });
   it('should throw NotInitializedError when metaFile does not exists', () => {
     stateManager_mock.metaFileExists.mockImplementation(() => false);
-    expect(() => getProjectMeta()).toThrow(
-      `No Amplify backend project files detected within this folder. Either initialize a new Amplify project or pull an existing project.
-- "amplify init" to initialize a new Amplify project
-- "amplify pull <app-id>" to pull your existing Amplify project. Find the <app-id> in the AWS Console or Amplify Admin UI.`,
-    );
+    expect(() => getProjectMeta()).toThrow(`
+      No Amplify backend project files detected within this folder. Either initialize a new Amplify project or pull an existing project.
+      - "amplify init" to initialize a new Amplify project
+      - "amplify pull <app-id>" to pull your existing Amplify project. Find the <app-id> in the AWS Console or Amplify Admin UI.
+      `);
   });
 });