+++ packages/amplify-e2e-core/src/index.ts
@@ -65,7 +65,7 @@ export async function createNewProjectDir(
   } while (fs.existsSync(projectDir));
 
   fs.ensureDirSync(projectDir);
-
+  console.log(projectDir);
   return projectDir;
 }