+++ packages/amplify-e2e-core/src/utils/sdk-calls.ts
@@ -14,6 +14,7 @@ import {
   AmplifyBackend,
   IAM,
   SSM,
+  Location,
 } from 'aws-sdk';
 import * as path from 'path';
 import _ from 'lodash';
@@ -349,3 +350,17 @@ export const getSSMParameters = async (region: string, appId: string, envName: s
     })
     .promise();
 };
+//Amazon location service calls
+export const getMap = async (mapName: string, region: string) => {
+  const service = new Location({region});
+  return await service.describeMap({
+    MapName: mapName
+  }).promise()
+}
+
+export const getPlaceIndex = async (placeIndexName: string, region: string) => {
+  const service = new Location({region});
+  return await service.describePlaceIndex({
+    IndexName: placeIndexName
+  }).promise()
+}