+++ packages/amplify-go-function-runtime-provider/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-go-function-runtime-provider@1.9.3...amplify-go-function-runtime-provider@1.9.4) (2021-09-18)
+
+**Note:** Version bump only for package amplify-go-function-runtime-provider
+
+
+
+
+
+## [1.9.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-go-function-runtime-provider@1.9.2...amplify-go-function-runtime-provider@1.9.3) (2021-09-14)
+
+**Note:** Version bump only for package amplify-go-function-runtime-provider
+
+
+
+
+
 ## [1.9.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-go-function-runtime-provider@1.9.1...amplify-go-function-runtime-provider@1.9.2) (2021-09-09)
 
 **Note:** Version bump only for package amplify-go-function-runtime-provider