+++ packages/graphql-dynamodb-transformer/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.22.22](https://github.com/aws-amplify/amplify-cli/compare/graphql-dynamodb-transformer@6.22.21...graphql-dynamodb-transformer@6.22.22) (2021-09-18)
+
+**Note:** Version bump only for package graphql-dynamodb-transformer
+
+
+
+
+
+## [6.22.21](https://github.com/aws-amplify/amplify-cli/compare/graphql-dynamodb-transformer@6.22.20...graphql-dynamodb-transformer@6.22.21) (2021-09-14)
+
+**Note:** Version bump only for package graphql-dynamodb-transformer
+
+
+
+
+
 ## [6.22.20](https://github.com/aws-amplify/amplify-cli/compare/graphql-dynamodb-transformer@6.22.19...graphql-dynamodb-transformer@6.22.20) (2021-09-09)
 
 **Note:** Version bump only for package graphql-dynamodb-transformer